etl/requirements.txt
psycopg2-binary	== 2.9.3
elasticsearch == 7.17.6
flake8==4.0
python-dotenv==0.20.0
pydantic==1.10.2